requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
# backports.zoneinfo==0.2.1
backports.zoneinfo;python_version<"3.9"
# Django==4.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
Django>=3.0,<4.0
# psycopg2>=2.8

psycopg2-binary==2.9.3
channels==3.0.4
django-channels-graphql-ws==0.9.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
Pillow==9.2.0
graphene-file-upload==1.3.0
graphene-validator==0.0.1
validator-collection==1.5.0
Jinja2==3.1.2
django-graphql-playground==1.0
requests==2.28.1
celery==5.2.7
gunicorn==20.1.0
python-decouple==3.6